reorganize functions and libraries
